Warning: I tried to checkout the most recent changes but the most recent it showed was my last push. I tried to commit and push it after but got an error saying a remote repository contained unmerged changes so I couldn't. Through a process of merging and resolving conflict I managed to get what I believe is the most recent changes uploaded to the branch and my changes combined. Everything is working right through testing, but if there's something weird or missing, it could be because of that.
Everything should be working. A couple things though: I used a lot of code similar to the book so any changes or optimizations are good, the checkout could use some css formatting (if you add an iphone and a fan to the cart then checkout, the price has too many decimals), and there's still the javascript validation on index.jsp, I'm not sure what to do about that. If we remove it the part on products.jsp that prints the error and in ProductsServlet that puts it in the request object should be removed.